requirements.txt
flask==3.0.0
python-telegram-bot==21.0
requests==2.31.0
langchain-google-genai==2.0.0
langchain-core==0.3.0
pydantic==2.5.0
jinja2==3.1.2